limit

+++ response.go
@@ -39,7 +39,7 @@ func (l *Limit) Max(max uint8){
 	l.Limit = min(l.Limit, max)
 	if l.Offset != 0 {
 		f := float64(l.Offset) / float64(l.Limit)
-		l.Offset = uint8(math.Round(f)) * l.Limit
+		l.Offset = uint32(math.Round(f)) * uint32(l.Limit)
 	}
 }
